suppression fini

+++ src/app/components/categorie-liste/categorie-liste.component.ts
@@ -9,6 +9,7 @@ import { CategorieService } from 'src/app/services/categorie.service';
 })
 export class CategorieListeComponent implements OnInit {
   categorie!:Categorie[]
+  message:boolean=false
 
   constructor(private categorieService:CategorieService) { }
 
@@ -22,7 +23,12 @@ export class CategorieListeComponent implements OnInit {
   }
   //suppression de categorie
   deleteCategorie(id:number){
+    let confirmation=confirm("Est-vous sur de vouloir supprimer ?"+id)
+    if(confirmation==false)return;
+    // this.categorie.splice(id,1)
     this.categorieService.deleteCategorie(id).subscribe((data)=>{
+      this.message=true;
+      this.ngOnInit();
     });
   }
   //modification de categore
@@ -32,6 +38,10 @@ export class CategorieListeComponent implements OnInit {
   //detail categorie
   detailCategrie(id:Number){
     alert("le detail de la categorie dont l'id est "+id)
+  }
+  //suppression de message de confirmation
+  removeMessage(){
+    this.message=false;
   }
 
 }
